arangopipe/requirements.txt
python-arango
pandas
PyYAML==5.4
sklearn2
yapf
autopep8
pre-commit
